// src/modules/warehouse/constants/warehouse.enums.ts
export enum WarehouseErrorKeyEnum {
  WAREHOUSE_WITH_THE_SAME_NAME_ALREADY_EXISTS = 'Warehouse with the same name already exists',
  FAILED_TO_UPDATE_DEFAULT_ADDRESS = 'Failed to update default address',
  FAILED_TO_CREATE_WAREHOUSE = 'Failed to create warehouse',
  FAILED_TO_FIND_WAREHOUSES = 'Failed to find warehouses',
  FAILED_TO_FIND_WAREHOUSE = 'Failed to find warehouse',
  WAREHOUSE_NOT_FOUND = 'Warehouse not found',
  FAILED_TO_UPDATE_ADDRESS = 'Failed to update address',
  FAILED_TO_UPDATE_WAREHOUSE = 'Failed to update warehouse',
  FAILED_TO_DELETE_WAREHOUSE = 'Failed to delete warehouse',
}